Resolve merge conflict

+++ src/SalesManager.java
@@ -16,9 +16,9 @@ public class SalesManager {
     }
 
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -27,9 +27,9 @@ public class SalesManager {
     }
 
     
-    public int getAverageSales() {
+    public long getAverageSales() {
         int sum = 0;
-        for (int sale : sales) {
+        for (long sale : sales) {
             sum += sale;
         }
         return (sum - min() - max()) / (sales.length - 2);
